+++ src/components/admin/blog/index.ts
@@ -1,3 +1,3 @@
-export { getBlogTableColumns } from "./BlogTableColumns";
+export { useBlogTableColumns } from "./BlogTableColumns";
 export { BlogForm } from "./BlogForm";
 export type { BlogFormData } from "./BlogForm";